Rename / move Json "tests" into prototypes area so they aren't confused with real tests.
