+++ TubeTests/EmptyTests.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace GlueTests
+namespace Tests
 {
     /// <summary>
     ///
