+++ TubeTests/TubeTests.csproj
@@ -61,7 +61,7 @@
     <Compile Include="FormatDurationFormatTests.cs" />
     <Compile Include="FormatDurationParseTests.cs" />
     <Compile Include="EmptyTests.cs" />
-    <Compile Include="JsonTests.cs" />
+    <Compile Include="Prototypes\JsonPrototype.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
